backend/requirements.txt
# Django Core
Django==5.2.9
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.3

# Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.1

# Image Processing
Pillow==10.4.0

# XML Processing
lxml==5.2.2
beautifulsoup4==4.12.3

# File Handling
python-magic==0.4.27

# API Documentation
drf-spectacular==0.27.2

# Utilities
python-slugify==8.0.4

# Production Server
gunicorn==22.0.0

# Development
pytest==8.2.2
pytest-django==4.8.0







